third_party/blink/web_tests/http/tests/devtools/console/console-last-result-expected.txt

Tests that console exposes last evaluation result as $_.


Running: testLastResult
1+1
2
$_
2

Running: testLastResultAfterConsoleClear
$_
undefined
